Update button UI to look closer to existing comps
Probably not exactly what we want, but at least it starts to put the
pieces in place. With the post button on the bottom we still get some
funky layout issues once there are multiple blocks. I believe this is
what we've been seeing with the Completers as well. Also updated the
icon for dismissing a block.

[Finishes: #143557767](https://www.pivotaltracker.com/story/show/143557767)